Added tests

+++ core/src/prediction/InstrumentalPredictionStrategy.cpp
@@ -243,6 +243,6 @@ std::vector<double> InstrumentalPredictionStrategy::compute_debiased_error(
     error_bias += (residual_loto - residual) * (residual_loto - residual);
   }
   
-  error_bias *= ((num_trees - 1) / num_trees);
+  error_bias *= ((double)(num_trees - 1) / num_trees);
   return { error_raw - error_bias };
 }
